+++ src/controller/src/controller/menu_bar/help_menu_controller.hpp
@@ -27,6 +27,7 @@ namespace controller
         ui::HelpMenu& _helpMenu;
 
        private slots:
+        void _onHelpPageRequested();
         void _onAboutRequested();
 
        public: